+++ classes/task/cron_task.php
@@ -163,10 +163,10 @@ class cron_task extends \core\task\scheduled_task {
         $postfix = '';
         if ($model instanceof restore_model) {
             $postfix = "\nIf the database restore did not finish, your site may be in an inconsistent state and will not work.".
-            ' You will need to re-install Moodle and repeat restore.';
+            ' You will need to re-install Moodle and repeat the restore process.';
         }
         $model->add_log('There was no activity for over ' . (constants::LOCK_TIMEOUT / 60) .
-            ' minutes. It is possible that cron was interrupted or timed out. '.
+            ' minutes. It is possible that the cron process was interrupted or timed out. '.
             'Operation is marked as failed, access to the site is now allowed.' . $postfix, constants::LOGLEVEL_ERROR);
         $model->set_status(constants::STATUS_FAILED)->save();
     }