+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the tool_vault plugin will be documented in this file.
 
+## Unreleased
+
+### Fixed
+- Fixed error when restoring data into tables with reserved words as fields if backup and
+  restore sites are on different database engines (mysql/postgres)
+
 ## [2.6] - 2024-10-13
 
 ### Added